Struct foliage::bevy_ecs::component::Components

pub struct Components { /* private fields */ }

Stores metadata associated with each kind of Component in a given World.

Implementations

impl Components

pub fn init_component<T>(&mut self, storages: &mut Storages) -> ComponentIdwhere T: Component,

Initializes a component of type T with this instance. If a component of this type has already been initialized, this will return the ID of the pre-existing component.

See also

Components::component_id()

Components::init_component_with_descriptor()

pub fn init_component_with_descriptor( &mut self, storages: &mut Storages, descriptor: ComponentDescriptor ) -> ComponentId

Initializes a component described by descriptor.

Note

If this method is called multiple times with identical descriptors, a distinct ComponentId will be created for each one.

See also

Components::component_id()

Components::init_component()

pub fn len(&self) -> usize

Returns the number of components registered with this instance.

pub fn is_empty(&self) -> bool

Returns true if there are no components registered with this instance. Otherwise, this returns false.

pub fn get_info(&self, id: ComponentId) -> Option<&ComponentInfo>

Gets the metadata associated with the given component.

This will return an incorrect result if id did not come from the same world as self. It may return None or a garbage value.

pub fn get_name(&self, id: ComponentId) -> Option<&str>

Returns the name associated with the given component.

This will return an incorrect result if id did not come from the same world as self. It may return None or a garbage value.

pub unsafe fn get_info_unchecked(&self, id: ComponentId) -> &ComponentInfo

Gets the metadata associated with the given component.

Safety

id must be a valid ComponentId

pub fn get_id(&self, type_id: TypeId) -> Option<ComponentId>

Type-erased equivalent of Components::component_id().

pub fn component_id<T>(&self) -> Option<ComponentId>where T: Component,

Returns the ComponentId of the given Component type T.

The returned ComponentId is specific to the Components instance it was retrieved from and should not be used with another Components instance.

Returns None if the Component type has not yet been initialized using Components::init_component().

use bevy_ecs::prelude::*;

let mut world = World::new();

#[derive(Component)]
struct ComponentA;

let component_a_id = world.init_component::<ComponentA>();

assert_eq!(component_a_id, world.components().component_id::<ComponentA>().unwrap())

See also

Components::get_id()

Components::resource_id()

World::component_id()

pub fn get_resource_id(&self, type_id: TypeId) -> Option<ComponentId>

Type-erased equivalent of Components::resource_id().

pub fn resource_id<T>(&self) -> Option<ComponentId>where T: Resource,

Returns the ComponentId of the given Resource type T.

The returned ComponentId is specific to the Components instance it was retrieved from and should not be used with another Components instance.

Returns None if the Resource type has not yet been initialized using Components::init_resource().

use bevy_ecs::prelude::*;

let mut world = World::new();

#[derive(Resource, Default)]
struct ResourceA;

let resource_a_id = world.init_resource::<ResourceA>();

assert_eq!(resource_a_id, world.components().resource_id::<ResourceA>().unwrap())

See also

Components::component_id()

Components::get_resource_id()

pub fn init_resource<T>(&mut self) -> ComponentIdwhere T: Resource,

Initializes a Resource of type T with this instance. If a resource of this type has already been initialized, this will return the ID of the pre-existing resource.

See also

Components::resource_id()

pub fn init_non_send<T>(&mut self) -> ComponentIdwhere T: Any,

Initializes a non-send resource of type T with this instance. If a resource of this type has already been initialized, this will return the ID of the pre-existing resource.

pub fn iter(&self) -> impl Iterator<Item = &ComponentInfo>

Gets an iterator over all components registered with this instance.