Struct foliage::bevy_ecs::prelude::Has

pub struct Has<T>(/* private fields */);

Returns a bool that describes if an entity has the component T.

This can be used in a Query if you want to know whether or not entities have the component T but don’t actually care about the component’s value.

Examples

fn food_entity_system(query: Query<(&Name, Has<IsHungry>) >) {
    for (name, is_hungry) in &query {
        if is_hungry{
            println!("{} would like some food.", name.name);
        } else {
            println!("{} has had sufficient.", name.name);
        }
    }
}

// Unlike `Option<&T>`, `Has<T>` is compatible with `&mut T`
// as it does not actually access any data.
fn alphabet_entity_system(mut alphas: Query<(&mut Alpha, Has<Beta>)>, mut betas: Query<(&mut Beta, Has<Alpha>)>) {
    for (mut alpha, has_beta) in alphas.iter_mut() {
        alpha.has_beta = has_beta;
    }
    for (mut beta, has_alpha) in betas.iter_mut() {
        beta.has_alpha = has_alpha;
    }
}