Struct foliage::bevy_ecs::system::EntityCommands

pub struct EntityCommands<'w, 's, 'a> { /* private fields */ }

A list of commands that will be run to modify an entity.

Implementations

impl<'w, 's, 'a> EntityCommands<'w, 's, 'a>

pub fn id(&self) -> Entity

Returns the Entity id of the entity.

Example

fn my_system(mut commands: Commands) {
    let entity_id = commands.spawn_empty().id();
}

pub fn insert(&mut self, bundle: impl Bundle) -> &mut EntityCommands<'w, 's, 'a>

Adds a Bundle of components to the entity.

This will overwrite any previous value(s) of the same component type.

Panics

The command will panic when applied if the associated entity does not exist.

To avoid a panic in this case, use the command Self::try_insert instead.

Example

#[derive(Component)]
struct Health(u32);
#[derive(Component)]
struct Strength(u32);
#[derive(Component)]
struct Defense(u32);

#[derive(Bundle)]
struct CombatBundle {
    health: Health,
    strength: Strength,
}

fn add_combat_stats_system(mut commands: Commands, player: Res<PlayerEntity>) {
    commands
        .entity(player.entity)
        // You can insert individual components:
        .insert(Defense(10))
        // You can also insert pre-defined bundles of components:
        .insert(CombatBundle {
            health: Health(100),
            strength: Strength(40),
        })
        // You can also insert tuples of components and bundles.
        // This is equivalent to the calls above:
        .insert((
            Defense(10),
            CombatBundle {
                health: Health(100),
                strength: Strength(40),
            },
        ));
}

pub fn try_insert( &mut self, bundle: impl Bundle ) -> &mut EntityCommands<'w, 's, 'a>

Tries to add a Bundle of components to the entity.

This will overwrite any previous value(s) of the same component type.

Note

Unlike Self::insert, this will not panic if the associated entity does not exist.

Example

#[derive(Component)]
struct Health(u32);
#[derive(Component)]
struct Strength(u32);
#[derive(Component)]
struct Defense(u32);

#[derive(Bundle)]
struct CombatBundle {
    health: Health,
    strength: Strength,
}

fn add_combat_stats_system(mut commands: Commands, player: Res<PlayerEntity>) {
  commands.entity(player.entity)
   // You can try_insert individual components:
    .try_insert(Defense(10))
     
   // You can also insert tuples of components:
    .try_insert(CombatBundle {
        health: Health(100),
        strength: Strength(40),
    });
    
   // Suppose this occurs in a parallel adjacent system or process
   commands.entity(player.entity)
     .despawn();

   commands.entity(player.entity)
   // This will not panic nor will it add the component
     .try_insert(Defense(5));
}

pub fn remove<T>(&mut self) -> &mut EntityCommands<'w, 's, 'a>where T: Bundle,

Removes a Bundle of components from the entity.

See EntityWorldMut::remove for more details.

Example

#[derive(Component)]
struct Health(u32);
#[derive(Component)]
struct Strength(u32);
#[derive(Component)]
struct Defense(u32);

#[derive(Bundle)]
struct CombatBundle {
    health: Health,
    strength: Strength,
}

fn remove_combat_stats_system(mut commands: Commands, player: Res<PlayerEntity>) {
    commands
        .entity(player.entity)
        // You can remove individual components:
        .remove::<Defense>()
        // You can also remove pre-defined Bundles of components:
        .remove::<CombatBundle>()
        // You can also remove tuples of components and bundles.
        // This is equivalent to the calls above:
        .remove::<(Defense, CombatBundle)>();
}

pub fn despawn(&mut self)

Despawns the entity.

See World::despawn for more details.

Panics

The command will panic when applied if the associated entity does not exist.

Example

fn remove_character_system(
    mut commands: Commands,
    character_to_remove: Res<CharacterToRemove>
)
{
    commands.entity(character_to_remove.entity).despawn();
}

pub fn add<C>(&mut self, command: C) -> &mut EntityCommands<'w, 's, 'a>where C: EntityCommand,

Pushes an EntityCommand to the queue, which will get executed for the current Entity.

Examples

commands
    .spawn_empty()
    // Closures with this signature implement `EntityCommand`.
    .add(|entity: EntityWorldMut| {
        println!("Executed an EntityCommand for {:?}", entity.id());
    });

pub fn log_components(&mut self)

Logs the components of the entity at the info level.

Panics

The command will panic when applied if the associated entity does not exist.

pub fn commands(&mut self) -> &mut Commands<'w, 's>

Returns the underlying Commands.