Struct foliage::bevy_ecs::system::NonSendMut

pub struct NonSendMut<'a, T>
where
    T: 'static + ?Sized,
{ /* private fields */ }

Unique borrow of a non-Send resource.

Only Send resources may be accessed with the ResMut SystemParam. In case that the resource does not implement Send, this SystemParam wrapper can be used. This will instruct the scheduler to instead run the system on the main thread so that it doesn’t send the resource over to another thread.

Panics

Panics when used as a SystemParameter if the resource does not exist.

Use Option<NonSendMut<T>> instead if the resource might not always exist.

Implementations

impl<'a, T> NonSendMut<'a, T>
where T: ?Sized,

pub fn into_inner(self) -> &'a mut T

Consume self and return a mutable reference to the contained value while marking self as “changed”.

pub fn reborrow(&mut self) -> Mut<'_, T>

Returns a Mut<> with a smaller lifetime. This is useful if you have &mut NonSendMut <T>, but you need a Mut<T>.

pub fn map_unchanged<U>(self, f: impl FnOnce(&mut T) -> &mut U) -> Mut<'a, U>
where U: ?Sized,

Maps to an inner value by applying a function to the contained reference, without flagging a change.

You should never modify the argument passed to the closure – if you want to modify the data without flagging a change, consider using DetectChangesMut::bypass_change_detection to make your intent explicit.

// When run, zeroes the translation of every entity.
fn reset_positions(mut transforms: Query<&mut Transform>) {
    for transform in &mut transforms {
        // We pinky promise not to modify `t` within the closure.
        // Breaking this promise will result in logic errors, but will never cause undefined behavior.
        let mut translation = transform.map_unchanged(|t| &mut t.translation);
        // Only reset the translation if it isn't already zero;
        translation.set_if_neq(Vec2::ZERO);
    }
}

pub fn as_deref_mut(&mut self) -> Mut<'_, <T as Deref>::Target>
where T: DerefMut,

Allows you access to the dereferenced value of this pointer without immediately triggering change detection.