foliage

Function any_component_removed

pub fn any_component_removed<T>(removals: RemovedComponents<'_, '_, T>) -> bool
where
    T: Component,

A SystemCondition-satisfying system that returns true if there are any entity with a component of the given type removed.