foliage

Function any_match_filter

pub fn any_match_filter<F>(query: Query<'_, '_, (), F>) -> bool
where
    F: QueryFilter,

A SystemCondition-satisfying system that returns true if there are any entities that match the given QueryFilter.

For a simple With<T> filter, this is equivalent to any_with_component::<T>().

To skip a system with a Query parameter if the query is empty, you may instead use Populated, if the query may match multiple entities, or Single, if it will only match one.