foliage

Struct Deferred

pub struct Deferred<'a, T>(/* private fields */)
where
    T: SystemBuffer;

A SystemParam that stores a buffer which gets applied to the World during ApplyDeferred. This is used internally by Commands to defer World mutations.

Examples

By using this type to defer mutations, you can avoid mutable World access within a system, which allows it to run in parallel with more systems.

Note that deferring mutations is not free, and should only be used if the gains in parallelization outweigh the time it takes to apply deferred mutations. In general, Deferred should only be used for mutations that are infrequent, or which otherwise take up a small portion of a system’s run-time.

// Tracks whether or not there is a threat the player should be aware of.
#[derive(Resource, Default)]
pub struct Alarm(bool);

#[derive(Component)]
pub struct Settlement {
    // ...
}

// A threat from inside the settlement.
#[derive(Component)]
pub struct Criminal;

// A threat from outside the settlement.
#[derive(Component)]
pub struct Monster;

use bevy_ecs::system::{Deferred, SystemBuffer, SystemMeta};

// Uses deferred mutations to allow signaling the alarm from multiple systems in parallel.
#[derive(Resource, Default)]
struct AlarmFlag(bool);

impl AlarmFlag {
    /// Sounds the alarm the next time buffers are applied via ApplyDeferred.
    pub fn flag(&mut self) {
        self.0 = true;
    }
}

impl SystemBuffer for AlarmFlag {
    // When `AlarmFlag` is used in a system, this function will get
    // called the next time buffers are applied via ApplyDeferred.
    fn queue(&mut self, system_meta: &SystemMeta, mut world: DeferredWorld) {
        if self.0 {
            world.resource_mut::<Alarm>().0 = true;
            self.0 = false;
        }
    }
}

// Sound the alarm if there are any criminals who pose a threat.
fn alert_criminal(
    settlement: Single<&Settlement>,
    criminals: Query<&Criminal>,
    mut alarm: Deferred<AlarmFlag>
) {
    for criminal in &criminals {
        // Only sound the alarm if the criminal is a threat.
        // For this example, assume that this check is expensive to run.
        // Since the majority of this system's run-time is dominated
        // by calling `is_threat()`, we defer sounding the alarm to
        // allow this system to run in parallel with other alarm systems.
        if criminal.is_threat(*settlement) {
            alarm.flag();
        }
    }
}

// Sound the alarm if there is a monster.
fn alert_monster(
    monsters: Query<&Monster>,
    mut alarm: ResMut<Alarm>
) {
    if monsters.iter().next().is_some() {
        // Since this system does nothing except for sounding the alarm,
        // it would be pointless to defer it, so we sound the alarm directly.
        alarm.0 = true;
    }
}

let mut world = World::new();
world.init_resource::<Alarm>();
world.spawn(Settlement {
    // ...
});

let mut schedule = Schedule::default();
// These two systems have no conflicts and will run in parallel.
schedule.add_systems((alert_criminal, alert_monster));

// There are no criminals or monsters, so the alarm is not sounded.
schedule.run(&mut world);
assert_eq!(world.resource::<Alarm>().0, false);

// Spawn a monster, which will cause the alarm to be sounded.
let m_id = world.spawn(Monster).id();
schedule.run(&mut world);
assert_eq!(world.resource::<Alarm>().0, true);

// Remove the monster and reset the alarm.
world.entity_mut(m_id).despawn();
world.resource_mut::<Alarm>().0 = false;

// Spawn a criminal, which will cause the alarm to be sounded.
world.spawn(Criminal);
schedule.run(&mut world);
assert_eq!(world.resource::<Alarm>().0, true);

Implementations

impl<T> Deferred<'_, T>
where T: SystemBuffer,

pub fn reborrow(&mut self) -> Deferred<'_, T>

Returns a Deferred<T> with a smaller lifetime. This is useful if you have &mut Deferred<T> but need Deferred<T>.