foliage

Struct Discard

pub struct Discard {
    pub entity: Entity,
}

Trigger emitted when a component is removed from an entity, regardless of whether or not it is later replaced.

Runs before the value is replaced, so you can still access the original component data. See ComponentHooks::on_discard for more information.

Fields
entity: Entity

The entity that held this component before it was discarded.