foliage

Struct EntityWorldMut

pub struct EntityWorldMut<'w> { /* private fields */ }

A mutable reference to a particular Entity, and the entire world.

This is essentially a performance-optimized (Entity, &mut World) tuple, which caches the EntityLocation to reduce duplicate lookups.

Since this type provides mutable access to the entire world, only one EntityWorldMut can exist at a time for a given world.

See also EntityMut, which allows disjoint mutable access to multiple entities at once. Unlike EntityMut, this type allows adding and removing components, and despawning the entity.

Invariants and Risk

An EntityWorldMut may point to a despawned entity. You can check this via is_despawned. Using an EntityWorldMut of a despawned entity may panic in some contexts, so read method documentation carefully.

Unless you have strong reason to assume these invariants, you should generally avoid keeping an EntityWorldMut to an entity that is potentially not spawned. For example, when inserting a component, that component insert may trigger an observer that despawns the entity. So, when you don’t have full knowledge of what commands may interact with this entity, do not further use this value without first checking is_despawned.

Implementations

impl<'w> EntityWorldMut<'w>

pub fn with_children( &mut self, func: impl FnOnce(&mut RelatedSpawner<'_, ChildOf>), ) -> &mut EntityWorldMut<'w>

Spawns children of this entity (with a ChildOf relationship) by taking a function that operates on a ChildSpawner. See also with_related.

pub fn add_children(&mut self, children: &[Entity]) -> &mut EntityWorldMut<'w>

Adds the given children to this entity. See also add_related.

pub fn clear_children(&mut self) -> &mut EntityWorldMut<'w>

👎Deprecated:
Use detach_all_children() instead

Removes all the children from this entity. See also detach_all_related

pub fn detach_all_children(&mut self) -> &mut EntityWorldMut<'w>

Removes all the parent-child relationships from this entity. To despawn the child entities, instead use EntityWorldMut::despawn_children. See also detach_all_related

pub fn insert_children( &mut self, index: usize, children: &[Entity], ) -> &mut EntityWorldMut<'w>

Insert children at specific index. See also insert_related.

pub fn insert_child( &mut self, index: usize, child: Entity, ) -> &mut EntityWorldMut<'w>

Insert child at specific index. See also insert_related.

pub fn add_child(&mut self, child: Entity) -> &mut EntityWorldMut<'w>

Adds the given child to this entity. See also add_related.

pub fn remove_children( &mut self, children: &[Entity], ) -> &mut EntityWorldMut<'w>

👎Deprecated:
Use detach_children() instead

Removes the relationship between this entity and the given entities.

pub fn detach_children( &mut self, children: &[Entity], ) -> &mut EntityWorldMut<'w>

Removes the parent-child relationship between this entity and the given entities. Does not despawn the children.

pub fn remove_child(&mut self, child: Entity) -> &mut EntityWorldMut<'w>

👎Deprecated:
Use detach_child() instead

Removes the relationship between this entity and the given entity.

pub fn detach_child(&mut self, child: Entity) -> &mut EntityWorldMut<'w>

Removes the parent-child relationship between this entity and the given entity. Does not despawn the child.

pub fn replace_children( &mut self, children: &[Entity], ) -> &mut EntityWorldMut<'w>

Replaces all the related children with a new set of children.

pub fn replace_children_with_difference( &mut self, entities_to_unrelate: &[Entity], entities_to_relate: &[Entity], newly_related_entities: &[Entity], ) -> &mut EntityWorldMut<'w>

Replaces all the related children with a new set of children.

Warning

Failing to maintain the functions invariants may lead to erratic engine behavior including random crashes. Refer to Self::replace_related_with_difference for a list of these invariants.

Panics

Panics when debug assertions are enabled if an invariant is broken and the command is executed.

pub fn with_child(&mut self, bundle: impl Bundle) -> &mut EntityWorldMut<'w>

Spawns the passed bundle and adds it to this entity as a child.

For efficient spawning of multiple children, use with_children.

impl<'w> EntityWorldMut<'w>

pub fn insert_reflect( &mut self, component: Box<dyn PartialReflect>, ) -> &mut EntityWorldMut<'w>

Adds the given boxed reflect component or bundle to the entity using the reflection data in AppTypeRegistry.

This will overwrite any previous component(s) of the same type.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

If AppTypeRegistry does not have the reflection data for the given Component or Bundle.

If the component or bundle data is invalid. See [PartialReflect::apply] for further details.

If AppTypeRegistry is not present in the World.

Note

Prefer to use the typed EntityWorldMut::insert if possible. Adding a reflected component is much slower.

pub fn insert_reflect_with_registry<T>( &mut self, component: Box<dyn PartialReflect>, ) -> &mut EntityWorldMut<'w>
where T: Resource + AsRef<TypeRegistry>,

Same as insert_reflect, but using the T resource as type registry instead of AppTypeRegistry.

This will overwrite any previous component(s) of the same type.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

If the given Resource does not have the reflection data for the given Component or Bundle.

If the component or bundle data is invalid. See [PartialReflect::apply] for further details.

If the given Resource is not present in the World.

pub fn remove_reflect( &mut self, component_type_path: Cow<'static, str>, ) -> &mut EntityWorldMut<'w>

Removes from the entity the component or bundle with the given type path registered in AppTypeRegistry.

If the type is a bundle, it will remove any components in that bundle regardless if the entity contains all the components.

Does nothing if the type is a component and the entity does not have a component of the same type, if the type is a bundle and the entity does not contain any of the components in the bundle, or if AppTypeRegistry does not contain the reflection data for the given component.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

If AppTypeRegistry is not present in the World.

Note

Prefer to use the typed EntityWorldMut::remove if possible. Removing a reflected component is much slower.

pub fn remove_reflect_with_registry<T>( &mut self, component_type_path: Cow<'static, str>, ) -> &mut EntityWorldMut<'w>
where T: Resource + AsRef<TypeRegistry>,

Same as remove_reflect, but using the T resource as type registry instead of AppTypeRegistry.

If the given type is a bundle, it will remove any components in that bundle regardless if the entity contains all the components.

Does nothing if the type is a component and the entity does not have a component of the same type, if the type is a bundle and the entity does not contain any of the components in the bundle, or if AppTypeRegistry does not contain the reflection data for the given component.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

If AppTypeRegistry is not present in the World.

pub fn take_reflect( &mut self, component_type_path: Cow<'static, str>, ) -> Option<Box<dyn Reflect>>

Takes from the entity the component or bundle with the given type path registered in AppTypeRegistry.

Does nothing and returns None if the type is a component and the entity does not have a component of the same type, if the type is a bundle and the entity does not contain every component in the bundle, or if AppTypeRegistry does not contain the reflection data for the given component.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

If AppTypeRegistry is not present in the World.

Note

Prefer to use the typed EntityWorldMut::take if possible. Taking a reflected component is much slower.

pub fn take_reflect_with_registry<T>( &mut self, component_type_path: Cow<'static, str>, ) -> Option<Box<dyn Reflect>>
where T: Resource + AsRef<TypeRegistry>,

Same as take_reflect, but using the T resource as type registry instead of AppTypeRegistry.

Does nothing and returns None if the type is a component and the entity does not have a component of the same type, if the type is a bundle and the entity does not contain every component in the bundle, or if AppTypeRegistry does not contain the reflection data for the given component.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

If AppTypeRegistry is not present in the World.

impl<'w> EntityWorldMut<'w>

pub fn with_related<R>( &mut self, bundle: impl Bundle, ) -> &mut EntityWorldMut<'w>
where R: Relationship,

Spawns a entity related to this entity (with the R relationship) by taking a bundle

pub fn with_related_entities<R>( &mut self, func: impl FnOnce(&mut RelatedSpawner<'_, R>), ) -> &mut EntityWorldMut<'w>
where R: Relationship,

Spawns entities related to this entity (with the R relationship) by taking a function that operates on a RelatedSpawner.

pub fn add_related<R>(&mut self, related: &[Entity]) -> &mut EntityWorldMut<'w>
where R: Relationship,

Relates the given entities to this entity with the relation R.

See add_one_related if you want relate only one entity.

pub fn detach_all_related<R>(&mut self) -> &mut EntityWorldMut<'w>
where R: Relationship,

Removes the relation R between this entity and all its related entities.

pub fn insert_related<R>( &mut self, index: usize, related: &[Entity], ) -> &mut EntityWorldMut<'w>
where R: Relationship, <<R as Relationship>::RelationshipTarget as RelationshipTarget>::Collection: OrderedRelationshipSourceCollection,

Relates the given entities to this entity with the relation R, starting at this particular index.

If the related has duplicates, a related entity will take the index of its last occurrence in related. If the indices go out of bounds, they will be clamped into bounds. This will not re-order existing related entities unless they are in related.

Example

use bevy_ecs::prelude::*;

let mut world = World::new();
let e0 = world.spawn_empty().id();
let e1 = world.spawn_empty().id();
let e2 = world.spawn_empty().id();
let e3 = world.spawn_empty().id();
let e4 = world.spawn_empty().id();

let mut main_entity = world.spawn_empty();
main_entity.add_related::<ChildOf>(&[e0, e1, e2, e2]);
main_entity.insert_related::<ChildOf>(1, &[e0, e3, e4, e4]);
let main_id = main_entity.id();

let relationship_source = main_entity.get::<Children>().unwrap().collection();
assert_eq!(relationship_source, &[e1, e0, e3, e2, e4]);

pub fn remove_related<R>( &mut self, related: &[Entity], ) -> &mut EntityWorldMut<'w>
where R: Relationship,

Removes the relation R between this entity and the given entities.

pub fn replace_related<R>( &mut self, related: &[Entity], ) -> &mut EntityWorldMut<'w>
where R: Relationship,

Replaces all the related entities with a new set of entities.

pub fn replace_related_with_difference<R>( &mut self, entities_to_unrelate: &[Entity], entities_to_relate: &[Entity], newly_related_entities: &[Entity], ) -> &mut EntityWorldMut<'w>
where R: Relationship,

Replaces all the related entities with a new set of entities.

This is a more efficient of Self::replace_related which doesn’t allocate. The passed in arguments must adhere to these invariants:

entities_to_unrelate: A slice of entities to remove from the relationship source. Entities need not be related to this entity, but must not appear in entities_to_relate

entities_to_relate: A slice of entities to relate to this entity. This must contain all entities that will remain related (i.e. not those in entities_to_unrelate) plus the newly related entities.

newly_related_entities: A subset of entities_to_relate containing only entities not already related to this entity.

Slices must not contain any duplicates

Warning

Violating these invariants may lead to panics, crashes or unpredictable engine behavior.

Panics

Panics when debug assertions are enabled and any invariants are broken.

pub fn add_one_related<R>(&mut self, entity: Entity) -> &mut EntityWorldMut<'w>
where R: Relationship,

Relates the given entity to this with the relation R.

See add_related if you want to relate more than one entity.

pub fn despawn_related<S>(&mut self) -> &mut EntityWorldMut<'w>
where S: RelationshipTarget,

Despawns entities that relate to this one via the given RelationshipTarget. This entity will not be despawned.

pub fn despawn_children(&mut self) -> &mut EntityWorldMut<'w>

Despawns the children of this entity. This entity will not be despawned.

This is a specialization of despawn_related, a more general method for despawning via relationships.

pub fn insert_recursive<S>( &mut self, bundle: impl Bundle + Clone, ) -> &mut EntityWorldMut<'w>
where S: RelationshipTarget,

Inserts a component or bundle of components into the entity and all related entities, traversing the relationship tracked in S in a breadth-first manner.

Warning

This method should only be called on relationships that form a tree-like structure. Any cycles will cause this method to loop infinitely.

pub fn remove_recursive<S, B>(&mut self) -> &mut EntityWorldMut<'w>
where S: RelationshipTarget, B: Bundle,

Removes a component or bundle of components of type B from the entity and all related entities, traversing the relationship tracked in S in a breadth-first manner.

Warning

This method should only be called on relationships that form a tree-like structure. Any cycles will cause this method to loop infinitely.

impl<'w> EntityWorldMut<'w>

pub fn into_readonly(self) -> EntityRef<'w>

Consumes self and returns read-only access to all of the entity’s components, with the world 'w lifetime.

pub fn as_readonly(&self) -> EntityRef<'_>

Gets read-only access to all of the entity’s components.

pub fn into_mutable(self) -> EntityMut<'w>

Consumes self and returns non-structural mutable access to all of the entity’s components, with the world 'w lifetime.

pub fn as_mutable(&mut self) -> EntityMut<'_>

Gets non-structural mutable access to all of the entity’s components.

pub fn id(&self) -> Entity

Returns the ID of the current entity.

pub fn try_location(&self) -> Option<EntityLocation>

Gets metadata indicating the location where the current entity is stored.

pub fn is_spawned(&self) -> bool

Returns if the entity is spawned or not.

pub fn try_archetype(&self) -> Option<&Archetype>

Returns the archetype that the current entity belongs to.

pub fn location(&self) -> EntityLocation

Gets metadata indicating the location where the current entity is stored.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn archetype(&self) -> &Archetype

Returns the archetype that the current entity belongs to.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn contains<T>(&self) -> bool
where T: Component,

Returns true if the current entity has a component of type T. Otherwise, this returns false.

Notes

If you do not know the concrete type of a component, consider using Self::contains_id or Self::contains_type_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn contains_id(&self, component_id: ComponentId) -> bool

Returns true if the current entity has a component identified by component_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you know the component’s TypeId but not its ComponentId, consider using Self::contains_type_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn contains_type_id(&self, type_id: TypeId) -> bool

Returns true if the current entity has a component with the type identified by type_id. Otherwise, this returns false.

Notes

If you know the concrete type of the component, you should prefer Self::contains.

If you have a ComponentId instead of a TypeId, consider using Self::contains_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get<T>(&self) -> Option<&T>
where T: Component,

Gets access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn components<Q>(&self) -> <Q as QueryData>::Item<'_, 'static>
where Q: ReadOnlyQueryData + ReleaseStateQueryData + SingleEntityQueryData,

Returns read-only components for the current entity that match the query Q.

Panics

If the entity does not have the components required by the query Q or if the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_components<Q>( &self, ) -> Result<<Q as QueryData>::Item<'_, 'static>, QueryAccessError>
where Q: ReadOnlyQueryData + ReleaseStateQueryData + SingleEntityQueryData,

Returns read-only components for the current entity that match the query Q, or None if the entity does not have the components required by the query Q.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn get_components_mut_unchecked<Q>( &mut self, ) -> Result<<Q as QueryData>::Item<'_, 'static>, QueryAccessError>
where Q: ReleaseStateQueryData + SingleEntityQueryData,

Returns components for the current entity that match the query Q, or None if the entity does not have the components required by the query Q.

Example

#[derive(Component)]
struct X(usize);
#[derive(Component)]
struct Y(usize);

let mut entity = world.spawn((X(0), Y(0)));
// Get mutable access to two components at once
// SAFETY: X and Y are different components
let (mut x, mut y) =
    unsafe { entity.get_components_mut_unchecked::<(&mut X, &mut Y)>() }.unwrap();
*x = X(1);
*y = Y(1);
// This would trigger undefined behavior, as the `&mut X`s would alias:
// entity.get_components_mut_unchecked::<(&mut X, &mut X)>();

Safety

It is the caller’s responsibility to ensure that the QueryData does not provide aliasing mutable references to the same component.

/// # See also

Self::get_components_mut for the safe version that performs aliasing checks

pub fn get_components_mut<Q>( &mut self, ) -> Result<<Q as QueryData>::Item<'_, 'static>, QueryAccessError>
where Q: ReleaseStateQueryData + SingleEntityQueryData,

Returns components for the current entity that match the query Q. In the case of conflicting QueryData, unregistered components, or missing components, this will return a QueryAccessError

Example

#[derive(Component)]
struct X(usize);
#[derive(Component)]
struct Y(usize);

let mut entity = world.spawn((X(0), Y(0))).into_mutable();
// Get mutable access to two components at once
// SAFETY: X and Y are different components
let (mut x, mut y) = entity.get_components_mut::<(&mut X, &mut Y)>().unwrap();

Note that this does a O(n^2) check that the QueryData does not conflict. If performance is a consideration you should use Self::get_components_mut_unchecked instead.

pub unsafe fn into_components_mut_unchecked<Q>( self, ) -> Result<<Q as QueryData>::Item<'w, 'static>, QueryAccessError>
where Q: ReleaseStateQueryData + SingleEntityQueryData,

Consumes self and returns components for the current entity that match the query Q for the world lifetime 'w, or None if the entity does not have the components required by the query Q.

Example

#[derive(Component)]
struct X(usize);
#[derive(Component)]
struct Y(usize);

let mut entity = world.spawn((X(0), Y(0)));
// Get mutable access to two components at once
// SAFETY: X and Y are different components
let (mut x, mut y) =
    unsafe { entity.into_components_mut_unchecked::<(&mut X, &mut Y)>() }.unwrap();
*x = X(1);
*y = Y(1);
// This would trigger undefined behavior, as the `&mut X`s would alias:
// entity.into_components_mut_unchecked::<(&mut X, &mut X)>();

Safety

It is the caller’s responsibility to ensure that the QueryData does not provide aliasing mutable references to the same component.

See also

Self::into_components_mut for the safe version that performs aliasing checks

pub fn into_components_mut<Q>( self, ) -> Result<<Q as QueryData>::Item<'w, 'static>, QueryAccessError>
where Q: ReleaseStateQueryData + SingleEntityQueryData,

Consumes self and returns components for the current entity that match the query Q for the world lifetime 'w, or None if the entity does not have the components required by the query Q.

The checks for aliasing mutable references may be expensive. If performance is a concern, consider making multiple calls to Self::get_mut. If that is not possible, consider using Self::into_components_mut_unchecked to skip the checks.

Panics

If the QueryData provides aliasing mutable references to the same component.

If the entity has been despawned while this EntityWorldMut is still alive.

Example

#[derive(Component)]
struct X(usize);
#[derive(Component)]
struct Y(usize);

let mut entity = world.spawn((X(0), Y(0)));
// Get mutable access to two components at once
let (mut x, mut y) = entity.into_components_mut::<(&mut X, &mut Y)>().unwrap();
*x = X(1);
*y = Y(1);

let mut entity = world.spawn((X(0)));
// This panics, as the `&mut X`s would alias:
entity.into_components_mut::<(&mut X, &mut X)>();

pub fn into_borrow<T>(self) -> Option<&'w T>
where T: Component,

Consumes self and gets access to the component of type T with the world 'w lifetime for the current entity. Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_ref<T>(&self) -> Option<Ref<'_, T>>
where T: Component,

Gets access to the component of type T for the current entity, including change detection information as a Ref.

Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn into_ref<T>(self) -> Option<Ref<'w, T>>
where T: Component,

Consumes self and gets access to the component of type T with the world 'w lifetime for the current entity, including change detection information as a Ref.

Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_mut<T>(&mut self) -> Option<Mut<'_, T>>
where T: Component<Mutability = Mutable>,

Gets mutable access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn modify_component<T, R>( &mut self, f: impl FnOnce(&mut T) -> R, ) -> Option<R>
where T: Component,

Temporarily removes a Component T from this Entity and runs the provided closure on it, returning the result if T was available. This will trigger the Remove and Discard component hooks without causing an archetype move.

This is most useful with immutable components, where removal and reinsertion is the only way to modify a value.

If you do not need to ensure the above hooks are triggered, and your component is mutable, prefer using get_mut.

Examples

#[derive(Component, PartialEq, Eq, Debug)]
#[component(immutable)]
struct Foo(bool);

entity.modify_component(|foo: &mut Foo| {
    foo.0 = true;
});

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn modify_component_by_id<R>( &mut self, component_id: ComponentId, f: impl for<'a> FnOnce(MutUntyped<'a>) -> R, ) -> Option<R>

Temporarily removes a Component T from this Entity and runs the provided closure on it, returning the result if T was available. This will trigger the Remove and Discard component hooks without causing an archetype move.

This is most useful with immutable components, where removal and reinsertion is the only way to modify a value.

If you do not need to ensure the above hooks are triggered, and your component is mutable, prefer using get_mut.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn get_mut_assume_mutable<T>(&mut self) -> Option<Mut<'_, T>>
where T: Component,

Gets mutable access to the component of type T for the current entity. Returns None if the entity does not have a component of type T.

Safety

T must be a mutable component

pub fn into_mut<T>(self) -> Option<Mut<'w, T>>
where T: Component<Mutability = Mutable>,

Consumes self and gets mutable access to the component of type T with the world 'w lifetime for the current entity. Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn into_mut_assume_mutable<T>(self) -> Option<Mut<'w, T>>
where T: Component,

Consumes self and gets mutable access to the component of type T with the world 'w lifetime for the current entity. Returns None if the entity does not have a component of type T.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

Safety

T must be a mutable component

pub fn resource<R>(&self) -> &R
where R: Resource,

Gets a reference to the resource of the given type

Panics

Panics if the resource does not exist. Use get_resource instead if you want to handle this case.

pub fn resource_mut<R>(&mut self) -> Mut<'_, R>
where R: Resource<Mutability = Mutable>,

Gets a mutable reference to the resource of the given type

Panics

Panics if the resource does not exist. Use get_resource_mut instead if you want to handle this case.

If you want to instead insert a value if the resource does not exist, use get_resource_or_insert_with.

pub fn get_resource<R>(&self) -> Option<&R>
where R: Resource,

Gets a reference to the resource of the given type if it exists

pub fn get_resource_mut<R>(&mut self) -> Option<Mut<'_, R>>
where R: Resource<Mutability = Mutable>,

Gets a mutable reference to the resource of the given type if it exists

pub fn resource_scope<R, U>( &mut self, f: impl FnOnce(&mut EntityWorldMut<'_>, Mut<'_, R>) -> U, ) -> U
where R: Resource,

Temporarily removes the requested resource from the World, runs custom user code, then re-adds the resource before returning.

Panics

Panics if the resource does not exist. Use try_resource_scope instead if you want to handle this case.

See World::resource_scope for further details.

pub fn try_resource_scope<R, U>( &mut self, f: impl FnOnce(&mut EntityWorldMut<'_>, Mut<'_, R>) -> U, ) -> Option<U>
where R: Resource,

Temporarily removes the requested resource from the World if it exists, runs custom user code, then re-adds the resource before returning. Returns None if the resource does not exist in the World.

See World::try_resource_scope for further details.

pub fn resource_entities(&self) -> &ResourceEntities

Retrieves this world’s ResourceEntities.

pub fn resource_entity<R>(&self) -> Option<Entity>
where R: Resource,

Retrieves the Entity associated with the resource of type R, if it exists.

pub fn get_change_ticks<T>(&self) -> Option<ComponentTicks>
where T: Component,

Retrieves the change ticks for the given component. This can be useful for implementing change detection in custom runtimes.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_changed_by<T>(&self) -> Option<MaybeLocation>
where T: Component,

Get the MaybeLocation from where the given Component was last changed from. This contains information regarding the last place (in code) that changed this component and can be useful for debugging. For more information, see Location, and enable the track_location feature.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_change_ticks_by_id( &self, component_id: ComponentId, ) -> Option<ComponentTicks>

Retrieves the change ticks for the given ComponentId. This can be useful for implementing change detection in custom runtimes.

You should prefer to use the typed API EntityWorldMut::get_change_ticks where possible and only use this in cases where the actual component types are not known at compile time.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_by_id<F>( &self, component_ids: F, ) -> Result<<F as DynamicComponentFetch>::Ref<'_>, EntityComponentError>
where F: DynamicComponentFetch,

Returns untyped read-only reference(s) to component(s) for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityWorldMut::get where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityWorldMut::get, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Examples

For examples on how to use this method, see EntityRef::get_by_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn into_borrow_by_id<F>( self, component_ids: F, ) -> Result<<F as DynamicComponentFetch>::Ref<'w>, EntityComponentError>
where F: DynamicComponentFetch,

Consumes self and returns untyped read-only reference(s) to component(s) with lifetime 'w for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityWorldMut::into_borrow where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityWorldMut::into_borrow, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Examples

For examples on how to use this method, see EntityRef::get_by_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn get_mut_by_id<F>( &mut self, component_ids: F, ) -> Result<<F as DynamicComponentFetch>::Mut<'_>, EntityComponentError>
where F: DynamicComponentFetch,

Returns untyped mutable reference(s) to component(s) for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityWorldMut::get_mut where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityWorldMut::get_mut, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Returns EntityComponentError::AliasedMutability if a component is requested multiple times.

Examples

For examples on how to use this method, see EntityMut::get_mut_by_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn get_mut_assume_mutable_by_id<F>( &mut self, component_ids: F, ) -> Result<<F as DynamicComponentFetch>::Mut<'_>, EntityComponentError>
where F: DynamicComponentFetch,

Returns untyped mutable reference(s) to component(s) for the current entity, based on the given ComponentIds. Assumes the given ComponentIds refer to mutable components.

You should prefer to use the typed API EntityWorldMut::get_mut_assume_mutable where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityWorldMut::get_mut_assume_mutable, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Returns EntityComponentError::AliasedMutability if a component is requested multiple times.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

Safety

It is the callers responsibility to ensure that

the provided ComponentIds must refer to mutable components.

pub fn into_mut_by_id<F>( self, component_ids: F, ) -> Result<<F as DynamicComponentFetch>::Mut<'w>, EntityComponentError>
where F: DynamicComponentFetch,

Consumes self and returns untyped mutable reference(s) to component(s) with lifetime 'w for the current entity, based on the given ComponentIds.

You should prefer to use the typed API EntityWorldMut::into_mut where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityWorldMut::into_mut, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Returns EntityComponentError::AliasedMutability if a component is requested multiple times.

Examples

For examples on how to use this method, see EntityMut::get_mut_by_id.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn into_mut_assume_mutable_by_id<F>( self, component_ids: F, ) -> Result<<F as DynamicComponentFetch>::Mut<'w>, EntityComponentError>
where F: DynamicComponentFetch,

Consumes self and returns untyped mutable reference(s) to component(s) with lifetime 'w for the current entity, based on the given ComponentIds. Assumes the given ComponentIds refer to mutable components.

You should prefer to use the typed API EntityWorldMut::into_mut_assume_mutable where possible and only use this in cases where the actual component types are not known at compile time.

Unlike EntityWorldMut::into_mut_assume_mutable, this returns untyped reference(s) to component(s), and it’s the job of the caller to ensure the correct type(s) are dereferenced (if necessary).

Errors

Returns EntityComponentError::MissingComponent if the entity does not have a component.

Returns EntityComponentError::AliasedMutability if a component is requested multiple times.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

Safety

It is the callers responsibility to ensure that

the provided ComponentIds must refer to mutable components.

pub fn insert<T>(&mut self, bundle: T) -> &mut EntityWorldMut<'w>
where T: Bundle,

Adds a Bundle of components to the entity.

This will overwrite any previous value(s) of the same component type.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn insert_with_relationship_hook_mode<T>( &mut self, bundle: T, relationship_hook_mode: RelationshipHookMode, ) -> &mut EntityWorldMut<'w>
where T: Bundle,

Adds a Bundle of components to the entity. Relationship components in the bundle will follow the configuration in relationship_hook_mode.

This will overwrite any previous value(s) of the same component type.

Warning

This can easily break the integrity of relationships. This is intended to be used for cloning and spawning code internals, not most user-facing scenarios.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn insert_if_new<T>(&mut self, bundle: T) -> &mut EntityWorldMut<'w>
where T: Bundle,

Adds a Bundle of components to the entity without overwriting.

This will leave any previous value(s) of the same component type unchanged.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn insert_by_id( &mut self, component_id: ComponentId, component: OwningPtr<'_>, ) -> &mut EntityWorldMut<'w>

Inserts a dynamic Component into the entity.

This will overwrite any previous value(s) of the same component type.

You should prefer to use the typed API EntityWorldMut::insert where possible.

Safety

ComponentId must be from the same world as EntityWorldMut

OwningPtr must be a valid reference to the type represented by ComponentId

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub unsafe fn insert_by_ids<'a, I>( &mut self, component_ids: &[ComponentId], iter_components: I, ) -> &mut EntityWorldMut<'w>
where I: Iterator<Item = OwningPtr<'a>>,

Inserts a dynamic Bundle into the entity.

This will overwrite any previous value(s) of the same component type.

You should prefer to use the typed API EntityWorldMut::insert where possible. If your Bundle only has one component, use the cached API EntityWorldMut::insert_by_id.

If possible, pass a sorted slice of ComponentId to maximize caching potential.

Safety

Each ComponentId must be from the same world as EntityWorldMut

Each OwningPtr must be a valid reference to the type represented by ComponentId

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn take<T>(&mut self) -> Option<T>
where T: Bundle + BundleFromComponents,

Removes all components in the Bundle from the entity and returns their previous values.

Note: If the entity does not have every component in the bundle, this method will not remove any of them.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn remove<T>(&mut self) -> &mut EntityWorldMut<'w>
where T: Bundle,

Removes any components in the Bundle from the entity.

See EntityCommands::remove for more details.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn remove_with_requires<T>(&mut self) -> &mut EntityWorldMut<'w>
where T: Bundle,

Removes all components in the Bundle and remove all required components for each component in the bundle

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn retain<T>(&mut self) -> &mut EntityWorldMut<'w>
where T: Bundle,

Removes any components except those in the Bundle (and its Required Components) from the entity.

See EntityCommands::retain for more details.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn remove_by_id( &mut self, component_id: ComponentId, ) -> &mut EntityWorldMut<'w>

Removes a dynamic Component from the entity if it exists.

You should prefer to use the typed API EntityWorldMut::remove where possible.

Panics

Panics if the provided ComponentId does not exist in the World or if the entity has been despawned while this EntityWorldMut is still alive.

pub fn remove_by_ids( &mut self, component_ids: &[ComponentId], ) -> &mut EntityWorldMut<'w>

Removes a dynamic bundle from the entity if it exists.

You should prefer to use the typed API EntityWorldMut::remove where possible.

Panics

Panics if any of the provided ComponentIds do not exist in the World or if the entity has been despawned while this EntityWorldMut is still alive.

pub fn clear(&mut self) -> &mut EntityWorldMut<'w>

Removes all components associated with the entity.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn despawn_no_free(self) -> Entity

Despawns the entity without freeing it to the allocator. This returns the new Entity, which you must manage. Note that this still increases the generation to differentiate different spawns of the same row.

Additionally, keep in mind the limitations documented in the type-level docs. Unless you have full knowledge of this EntityWorldMut’s lifetime, you may not assume that nothing else has taken responsibility of this Entity. If you are not careful, this could cause a double free.

This may be later spawn_at. See World::despawn_no_free for details and usage examples.

pub fn template_context<T>( &mut self, func: impl FnOnce(&mut TemplateContext<'_, '_>) -> Result<T, BevyError>, ) -> Result<T, BevyError>

Creates a new TemplateContext for this entity and passes it into the given func.

pub fn build_template<T>( &mut self, template: &T, ) -> Result<<T as Template>::Output, BevyError>
where T: Template,

Builds the given template using a TemplateContext generated for this entity.

pub fn despawn(self)

Despawns the current entity.

See World::despawn for more details.

Note

This will also despawn any Children entities, and any other RelationshipTarget that is configured to despawn descendants. This results in “recursive despawn” behavior.

pub fn flush(self) -> Entity

Ensures any commands triggered by the actions of Self are applied, equivalent to World::flush

pub fn world(&self) -> &World

Gets read-only access to the world that the current entity belongs to.

pub unsafe fn world_mut(&mut self) -> &mut World

Returns this entity’s world.

See EntityWorldMut::world_scope or EntityWorldMut::into_world_mut for a safe alternative.

Safety

Caller must not modify the world in a way that changes the current entity’s location If the caller does do something that could change the location, self.update_location() must be called before using any other methods on this EntityWorldMut.

pub fn into_world_mut(self) -> &'w mut World

Returns this entity’s World, consuming itself.

pub fn world_scope<U>(&mut self, f: impl FnOnce(&mut World) -> U) -> U

Gives mutable access to this entity’s World in a temporary scope. This is a safe alternative to using EntityWorldMut::world_mut.

Examples

#[derive(Resource, Default, Clone, Copy)]
struct R(u32);

// This closure gives us temporary access to the world.
let new_r = entity.world_scope(|world: &mut World| {
    // Mutate the world while we have access to it.
    let mut r = world.resource_mut::<R>();
    r.0 += 1;

    // Return a value from the world before giving it back to the `EntityWorldMut`.
    *r
});

pub fn update_location(&mut self)

Updates the internal entity location to match the current location in the internal World.

This is only required when using the unsafe function EntityWorldMut::world_mut, which enables the location to change.

Note that if the entity is not spawned for any reason, this will have a location of None, leading some methods to panic.

pub fn is_despawned(&self) -> bool

Returns if the entity has been despawned.

Normally it shouldn’t be needed to explicitly check if the entity has been despawned between commands as this shouldn’t happen. However, for some special cases where it is known that a hook or an observer might despawn the entity while a EntityWorldMut reference is still held, this method can be used to check if the entity is still alive to avoid panicking when calling further methods.

pub fn entry<'a, T>(&'a mut self) -> ComponentEntry<'w, 'a, T>
where T: Component,

Gets an Entry into the world for this entity and component for in-place manipulation.

The type parameter specifies which component to get.

Examples

#[derive(Component, Default, Clone, Copy, Debug, PartialEq)]
struct Comp(u32);

let mut entity = world.spawn_empty();
entity.entry().or_insert_with(|| Comp(4));
assert_eq!(world.query::<&Comp>().single(&world).unwrap().0, 4);

entity.entry::<Comp>().and_modify(|mut c| c.0 += 1);
assert_eq!(world.query::<&Comp>().single(&world).unwrap().0, 5);

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

pub fn observe<M>( &mut self, observer: impl IntoEntityObserver<M>, ) -> &mut EntityWorldMut<'w>

Creates an Observer watching for an EntityEvent of type E whose EntityEvent::event_target targets this entity.

Panics

If the entity has been despawned while this EntityWorldMut is still alive.

Panics if the given system is an exclusive system.

pub fn clone_with_opt_out( &mut self, target: Entity, config: impl FnOnce(&mut EntityClonerBuilder<'_, OptOut>) + Send + Sync + 'static, ) -> &mut EntityWorldMut<'w>

Clones parts of an entity (components, observers, etc.) onto another entity, configured through EntityClonerBuilder.

The other entity will receive all the components of the original that implement Clone or Reflect except those that are denied in the config.

Example

// Clone all components except ComponentA onto the target.
world.entity_mut(entity).clone_with_opt_out(target, |builder| {
    builder.deny::<ComponentA>();
});

See EntityClonerBuilder<OptOut> for more options.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

If the target entity does not exist.

pub fn clone_with_opt_in( &mut self, target: Entity, config: impl FnOnce(&mut EntityClonerBuilder<'_, OptIn>) + Send + Sync + 'static, ) -> &mut EntityWorldMut<'w>

Clones parts of an entity (components, observers, etc.) onto another entity, configured through EntityClonerBuilder.

The other entity will receive only the components of the original that implement Clone or Reflect and are allowed in the config.

Example

// Clone only ComponentA onto the target.
world.entity_mut(entity).clone_with_opt_in(target, |builder| {
    builder.allow::<ComponentA>();
});

See EntityClonerBuilder<OptIn> for more options.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

If the target entity does not exist.

pub fn clone_and_spawn(&mut self) -> Entity

Spawns a clone of this entity and returns the Entity of the clone.

The clone will receive all the components of the original that implement Clone or Reflect.

To configure cloning behavior (such as only cloning certain components), use EntityWorldMut::clone_and_spawn_with_opt_out/ opt_in.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

pub fn clone_and_spawn_with_opt_out( &mut self, config: impl FnOnce(&mut EntityClonerBuilder<'_, OptOut>) + Send + Sync + 'static, ) -> Entity

Spawns a clone of this entity and allows configuring cloning behavior using EntityClonerBuilder, returning the Entity of the clone.

The clone will receive all the components of the original that implement Clone or Reflect except those that are denied in the config.

Example

// Create a clone of an entity but without ComponentA.
let entity_clone = world.entity_mut(entity).clone_and_spawn_with_opt_out(|builder| {
    builder.deny::<ComponentA>();
});

See EntityClonerBuilder<OptOut> for more options.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

pub fn clone_and_spawn_with_opt_in( &mut self, config: impl FnOnce(&mut EntityClonerBuilder<'_, OptIn>) + Send + Sync + 'static, ) -> Entity

Spawns a clone of this entity and allows configuring cloning behavior using EntityClonerBuilder, returning the Entity of the clone.

The clone will receive only the components of the original that implement Clone or Reflect and are allowed in the config.

Example

// Create a clone of an entity but only with ComponentA.
let entity_clone = world.entity_mut(entity).clone_and_spawn_with_opt_in(|builder| {
    builder.allow::<ComponentA>();
});

See EntityClonerBuilder<OptIn> for more options.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

pub fn clone_components<B>(&mut self, target: Entity) -> &mut EntityWorldMut<'w>
where B: Bundle,

Clones the specified components of this entity and inserts them into another entity.

Components can only be cloned if they implement Clone or Reflect.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

If the target entity does not exist.

pub fn move_components<B>(&mut self, target: Entity) -> &mut EntityWorldMut<'w>
where B: Bundle,

Clones the specified components of this entity and inserts them into another entity, then removes the components from this entity.

Components can only be cloned if they implement Clone or Reflect.

Panics

If this entity has been despawned while this EntityWorldMut is still alive.

If the target entity does not exist.

pub fn spawned_by(&self) -> MaybeLocation

Returns the source code location from which this entity has last been spawned.

pub fn spawn_tick(&self) -> Tick

Returns the Tick at which this entity has last been spawned.

pub fn reborrow_scope<U>( &mut self, f: impl FnOnce(EntityWorldMut<'_>) -> U, ) -> U

Reborrows this entity in a temporary scope. This is useful for executing a function that requires a EntityWorldMut but you do not want to move out the entity ownership.

pub fn trigger<'t, E>( &mut self, event_fn: impl FnOnce(Entity) -> E, ) -> &mut EntityWorldMut<'w>
where E: EntityEvent, <E as Event>::Trigger<'t>: Default,

Passes the current entity into the given function, and triggers the EntityEvent returned by that function. See EntityCommands::trigger for usage examples