foliage

Struct Messages

pub struct Messages<M>
where
    M: Message,
{ /* private fields */ }

A message collection that represents the messages that occurred within the last two Messages::update calls. Messages can be written to using a MessageWriter and are typically cheaply read using a MessageReader.

Each message can be consumed by multiple systems, in parallel, with consumption tracked by the MessageReader on a per-system basis.

If no ordering is applied between writing and reading systems, there is a risk of a race condition. This means that whether the messages arrive before or after the next Messages::update is unpredictable.

This collection is meant to be paired with a system that calls Messages::update exactly once per update/frame.

message_update_system is a system that does this, typically initialized automatically using add_message. MessageReaders are expected to read messages from this collection at least once per loop/frame. Messages will persist across a single frame boundary and so ordering of message producers and consumers is not critical (although poorly-planned ordering may cause accumulating lag). If messages are not handled by the end of the frame after they are updated, they will be dropped silently.

Example

use bevy_ecs::message::{Message, Messages};

#[derive(Message)]
struct MyMessage {
    value: usize
}

// setup
let mut messages = Messages::<MyMessage>::default();
let mut cursor = messages.get_cursor();

// run this once per update/frame
messages.update();

// somewhere else: write a message
messages.write(MyMessage { value: 1 });

// somewhere else: read the messages
for message in cursor.read(&messages) {
    assert_eq!(message.value, 1)
}

// messages are only processed once per reader
assert_eq!(cursor.read(&messages).count(), 0);

Details

Messages is implemented using a variation of a double buffer strategy. Each call to update swaps buffers and clears out the oldest one.

MessageReaders will read messages from both buffers.

MessageReaders that read at least once per update will never drop messages.

MessageReaders that read once within two updates might still receive some messages

MessageReaders that read after two updates are guaranteed to drop all messages that occurred before those updates.

The buffers in Messages will grow indefinitely if update is never called.

An alternative call pattern would be to call update manually across frames to control when messages are cleared. This complicates consumption and risks ever-expanding memory usage if not cleaned up, but can be done by adding your message as a resource instead of using add_message.

Example usage. Example usage standalone.

Implementations

impl<M> Messages<M>
where M: Message,

pub fn oldest_message_count(&self) -> usize

Returns the index of the oldest message stored in the message buffer.

pub fn write(&mut self, message: M) -> MessageId<M>

Writes an message to the current message buffer. MessageReaders can then read the message. This method returns the ID of the written message.

pub fn write_batch( &mut self, messages: impl IntoIterator<Item = M>, ) -> WriteBatchIds<M>

Writes a list of messages all at once, which can later be read by MessageReaders. This is more efficient than writing each message individually. This method returns the IDs of the written messages.

pub fn write_default(&mut self) -> MessageId<M>
where M: Default,

Writes the default value of the message. Useful when the message is an empty struct. This method returns the ID of the written message.

pub fn get_cursor(&self) -> MessageCursor<M>

Gets a new MessageCursor. This will include all messages already in the message buffers.

pub fn get_cursor_current(&self) -> MessageCursor<M>

Gets a new MessageCursor. This will ignore all messages already in the message buffers. It will read all future messages.

pub fn update(&mut self)

Swaps the message buffers and clears the oldest message buffer. In general, this should be called once per frame/update.

If you need access to the messages that were removed, consider using Messages::update_drain.

pub fn update_drain(&mut self) -> impl Iterator<Item = M>

Swaps the message buffers and drains the oldest message buffer, returning an iterator of all messages that were removed. In general, this should be called once per frame/update.

If you do not need to take ownership of the removed messages, use Messages::update instead.

pub fn clear(&mut self)

Removes all messages.

pub fn len(&self) -> usize

Returns the number of messages currently stored in the message buffer.

pub fn is_empty(&self) -> bool

Returns true if there are no messages currently stored in the message buffer.

pub fn drain(&mut self) -> impl Iterator<Item = M>

Creates a draining iterator that removes all messages.

pub fn iter_current_update_messages(&self) -> impl ExactSizeIterator

Iterates over messages that happened since the last “update” call. WARNING: You probably don’t want to use this call. In most cases you should use an MessageReader. You should only use this if you know you only need to consume messages between the last update() call and your call to iter_current_update_messages. If messages happen outside that window, they will not be handled. For example, any messages that happen after this call and before the next update() call will be dropped.

pub fn get_message(&self, id: usize) -> Option<(&M, MessageId<M>)>

Get a specific message by id if it still exists in the messages buffer.