foliage

Struct Mut

pub struct Mut<'w, T>
where
    T: ?Sized,
{ /* private fields */ }

Unique mutable borrow of an entity’s component or of a resource.

This can be used in queries to access change detection from immutable query methods, as opposed to &mut T which only provides access to change detection from mutable query methods.

#[derive(Component, Clone, Debug)]
struct Name(String);

#[derive(Component, Clone, Copy, Debug)]
struct Health(f32);

fn my_system(mut query: Query<(Mut<Name>, &mut Health)>) {
    // Mutable access provides change detection information for both parameters:
    // - `name` has type `Mut<Name>`
    // - `health` has type `Mut<Health>`
    for (name, health) in query.iter_mut() {
        println!("Name: {:?} (last changed {:?})", name, name.last_changed());
        println!("Health: {:?} (last changed: {:?})", health, health.last_changed());
    }

    // Immutable access only provides change detection for `Name`:
    // - `name` has type `Ref<Name>`
    // - `health` has type `&Health`
    for (name, health) in query.iter() {
        println!("Name: {:?} (last changed {:?})", name, name.last_changed());
        println!("Health: {:?}", health);
    }
}

Implementations

impl<'w, T> Mut<'w, T>
where T: ?Sized,

pub fn new( value: &'w mut T, added: &'w mut Tick, last_changed: &'w mut Tick, last_run: Tick, this_run: Tick, caller: MaybeLocation<&'w mut &'static Location<'static>>, ) -> Mut<'w, T>

Creates a new change-detection enabled smart pointer. In almost all cases you do not need to call this method manually, as instances of Mut will be created by engine-internal code.

Many use-cases of this method would be better served by Mut::map_unchanged or Mut::reborrow.

value - The value wrapped by this smart pointer.

added - A Tick that stores the tick when the wrapped value was created.

last_changed - A Tick that stores the last time the wrapped value was changed. This will be updated to the value of change_tick if the returned smart pointer is modified.

last_run - A Tick, occurring before this_run, which is used as a reference to determine whether the wrapped value is newly added or changed.

this_run - A Tick corresponding to the current point in time – “now”.

pub fn set_ticks(&mut self, last_run: Tick, this_run: Tick)

Overwrite the last_run and this_run tick that are used for change detection.

This is an advanced feature. Muts are usually created by engine-internal code and consumed by end-user code.

impl<'w, T> Mut<'w, T>
where T: ?Sized,

pub fn into_inner(self) -> &'w mut T

Consume self and return a mutable reference to the contained value while marking self as “changed”.

pub fn reborrow(&mut self) -> Mut<'_, T>

Returns a Mut<> with a smaller lifetime. This is useful if you have &mut Mut <T>, but you need a Mut<T>.

pub fn map_unchanged<U>(self, f: impl FnOnce(&mut T) -> &mut U) -> Mut<'w, U>
where U: ?Sized,

Maps to an inner value by applying a function to the contained reference, without flagging a change.

You should never modify the argument passed to the closure – if you want to modify the data without flagging a change, consider using DetectChangesMut::bypass_change_detection to make your intent explicit.

// When run, zeroes the translation of every entity.
fn reset_positions(mut transforms: Query<&mut Transform>) {
    for transform in &mut transforms {
        // We pinky promise not to modify `t` within the closure.
        // Breaking this promise will result in logic errors, but will never cause undefined behavior.
        let mut translation = transform.map_unchanged(|t| &mut t.translation);
        // Only reset the translation if it isn't already zero;
        translation.set_if_neq(Vec2::ZERO);
    }
}

pub fn filter_map_unchanged<U>( self, f: impl FnOnce(&mut T) -> Option<&mut U>, ) -> Option<Mut<'w, U>>
where U: ?Sized,

Optionally maps to an inner value by applying a function to the contained reference. This is useful in a situation where you need to convert a Mut<T> to a Mut<U>, but only if T contains U.

As with map_unchanged, you should never modify the argument passed to the closure.

pub fn try_map_unchanged<U, E>( self, f: impl FnOnce(&mut T) -> Result<&mut U, E>, ) -> Result<Mut<'w, U>, E>
where U: ?Sized,

Optionally maps to an inner value by applying a function to the contained reference, returns an error on failure. This is useful in a situation where you need to convert a Mut<T> to a Mut<U>, but only if T contains U.

As with map_unchanged, you should never modify the argument passed to the closure.

pub fn as_deref_mut(&mut self) -> Mut<'_, <T as Deref>::Target>
where T: DerefMut,

Allows you access to the dereferenced value of this pointer without immediately triggering change detection.