foliage

Struct Or

pub struct Or<T>(/* private fields */);

A filter that tests if any of the given filters apply.

This is useful for example if a system with multiple components in a query only wants to run when one or more of the components have changed.

The And equivalent to this filter is a tuple testing that all the contained filters apply instead.

Examples

fn print_cool_entity_system(query: Query<Entity, Or<(Changed<Color>, Changed<Node>)>>) {
    for entity in &query {
        println!("Entity {} got a new style or color", entity);
    }
}