foliage

Struct ParamSet

pub struct ParamSet<'w, 's, T>
where
    T: SystemParam,
{ /* private fields */ }

A collection of potentially conflicting SystemParams allowed by disjoint access.

Allows systems to safely access and interact with up to 8 mutually exclusive SystemParams, such as two queries that reference the same mutable data or an event reader and writer of the same type.

Each individual SystemParam can be accessed by using the functions p0(), p1(), …, p7(), according to the order they are defined in the ParamSet. This ensures that there’s either only one mutable reference to a parameter at a time or any number of immutable references.

Examples

The following system mutably accesses the same component two times, which is not allowed due to rust’s mutability rules.

// This will panic at runtime when the system gets initialized.
fn bad_system(
    mut enemies: Query<&mut Health, With<Enemy>>,
    mut allies: Query<&mut Health, With<Ally>>,
) {
    // ...
}

Conflicting SystemParams like these can be placed in a ParamSet, which leverages the borrow checker to ensure that only one of the contained parameters are accessed at a given time.

// Given the following system
fn fancy_system(
    mut set: ParamSet<(
        Query<&mut Health, With<Enemy>>,
        Query<&mut Health, With<Ally>>,
    )>
) {
    // This will access the first `SystemParam`.
    for mut health in set.p0().iter_mut() {
        // Do your fancy stuff here...
    }

    // The second `SystemParam`.
    // This would fail to compile if the previous parameter was still borrowed.
    for mut health in set.p1().iter_mut() {
        // Do even fancier stuff here...
    }
}

Of course, ParamSets can be used with any kind of SystemParam, not just queries.

fn message_system(
    mut set: ParamSet<(
        // PROBLEM: `MessageReader` and `MessageWriter` cannot be used together normally,
        // because they both need access to the same message queue.
        // SOLUTION: `ParamSet` allows these conflicting parameters to be used safely
        // by ensuring only one is accessed at a time.
        // Note that a better solution here is to use `MessageMutator`,
        // which both reads and writes messages with a single parameter.
        MessageReader<MyMessage>,
        MessageWriter<MyMessage>,
        // PROBLEM: `&World` needs read access to everything, which conflicts with
        // any mutable access in the same system.
        // SOLUTION: `ParamSet` ensures `&World` is only accessed when we're not
        // using the other mutable parameters.
        &World,
    )>,
) {
    for message in set.p0().read() {
        // ...
    }
    set.p1().write(MyMessage::new());

    let entities = set.p2().entities();
    // ...
}

Implementations

impl<'w, 's, P0> ParamSet<'w, 's, (P0,)>
where P0: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1> ParamSet<'w, 's, (P0, P1)>
where P0: SystemParam, P1: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1, P2> ParamSet<'w, 's, (P0, P1, P2)>
where P0: SystemParam, P1: SystemParam, P2: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p2<'a>(&'a mut self) -> <P2 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 2 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1, P2, P3> ParamSet<'w, 's, (P0, P1, P2, P3)>
where P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p2<'a>(&'a mut self) -> <P2 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 2 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p3<'a>(&'a mut self) -> <P3 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 3 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1, P2, P3, P4> ParamSet<'w, 's, (P0, P1, P2, P3, P4)>
where P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p2<'a>(&'a mut self) -> <P2 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 2 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p3<'a>(&'a mut self) -> <P3 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 3 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p4<'a>(&'a mut self) -> <P4 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 4 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1, P2, P3, P4, P5> ParamSet<'w, 's, (P0, P1, P2, P3, P4, P5)>
where P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p2<'a>(&'a mut self) -> <P2 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 2 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p3<'a>(&'a mut self) -> <P3 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 3 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p4<'a>(&'a mut self) -> <P4 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 4 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p5<'a>(&'a mut self) -> <P5 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 5 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1, P2, P3, P4, P5, P6> ParamSet<'w, 's, (P0, P1, P2, P3, P4, P5, P6)>
where P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p2<'a>(&'a mut self) -> <P2 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 2 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p3<'a>(&'a mut self) -> <P3 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 3 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p4<'a>(&'a mut self) -> <P4 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 4 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p5<'a>(&'a mut self) -> <P5 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 5 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p6<'a>(&'a mut self) -> <P6 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 6 in this ParamSet. No other parameters may be accessed while this one is active.

impl<'w, 's, P0, P1, P2, P3, P4, P5, P6, P7> ParamSet<'w, 's, (P0, P1, P2, P3, P4, P5, P6, P7)>
where P0: SystemParam, P1: SystemParam, P2: SystemParam, P3: SystemParam, P4: SystemParam, P5: SystemParam, P6: SystemParam, P7: SystemParam,

pub fn p0<'a>(&'a mut self) -> <P0 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 0 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p1<'a>(&'a mut self) -> <P1 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 1 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p2<'a>(&'a mut self) -> <P2 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 2 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p3<'a>(&'a mut self) -> <P3 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 3 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p4<'a>(&'a mut self) -> <P4 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 4 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p5<'a>(&'a mut self) -> <P5 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 5 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p6<'a>(&'a mut self) -> <P6 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 6 in this ParamSet. No other parameters may be accessed while this one is active.

pub fn p7<'a>(&'a mut self) -> <P7 as SystemParam>::Item<'a, 'a>

Gets exclusive access to the parameter at index 7 in this ParamSet. No other parameters may be accessed while this one is active.

impl<T> ParamSet<'_, '_, Vec<T>>
where T: SystemParam,

pub fn get_mut(&mut self, index: usize) -> <T as SystemParam>::Item<'_, '_>

Accesses the parameter at the given index. No other parameters may be accessed while this one is active.

pub fn for_each(&mut self, f: impl FnMut(<T as SystemParam>::Item<'_, '_>))

Calls a closure for each parameter in the set.