foliage

Struct Populated

pub struct Populated<'w, 's, D, F = ()>(/* private fields */)
where
    D: QueryData,
    F: QueryFilter;

System parameter that works very much like Query except it always contains at least one matching entity.

This SystemParam fails validation if no matching entities exist. This will cause the system to be skipped, according to the rules laid out in SystemParamValidationError.

Much like Query::is_empty the worst case runtime will be O(n) where n is the number of potential matches. This can be notably expensive for queries that rely on non-archetypal filters such as Added, Changed of Spawned which must individually check each query result for a match.

See Query for more details.

If the system doesn’t need to perform the query but should still be skipped if it is empty, you may use the any_with_component or any_match_filter run conditions.

Implementations

impl<'w, 's, D, F> Populated<'w, 's, D, F>
where D: QueryData, F: QueryFilter,

pub fn into_inner(self) -> Query<'w, 's, D, F>

Returns the inner item with ownership.