foliage

Struct Ref

pub struct Ref<'w, T>
where
    T: ?Sized,
{ /* private fields */ }

Shared borrow of an entity’s component with access to change detection. Similar to Mut but is immutable and so doesn’t require unique access.

Examples

These two systems produce the same output.

fn how_many_changed_1(query: Query<(), Changed<MyComponent>>) {
    println!("{} changed", query.iter().count());
}

fn how_many_changed_2(query: Query<Ref<MyComponent>>) {
    println!("{} changed", query.iter().filter(|c| c.is_changed()).count());
}

Implementations

impl<'w, T> Ref<'w, T>
where T: ?Sized,

pub fn into_inner(self) -> &'w T

Returns the reference wrapped by this type. The reference is allowed to outlive self, which makes this method more flexible than simply borrowing self.

pub fn map<U>(self, f: impl FnOnce(&T) -> &U) -> Ref<'w, U>
where U: ?Sized,

Map Ref to a different type using f.

This doesn’t do anything else than call f on the wrapped value. This is equivalent to Mut::map_unchanged.

pub fn new( value: &'w T, added: &'w Tick, changed: &'w Tick, last_run: Tick, this_run: Tick, caller: MaybeLocation<&'w &'static Location<'static>>, ) -> Ref<'w, T>

Create a new Ref using provided values.

This is an advanced feature, Refs are designed to be created by engine-internal code and consumed by end-user code.

value - The value wrapped by Ref.

added - A Tick that stores the tick when the wrapped value was created.

changed - A Tick that stores the last time the wrapped value was changed.

last_run - A Tick, occurring before this_run, which is used as a reference to determine whether the wrapped value is newly added or changed.

this_run - A Tick corresponding to the current point in time – “now”.

pub fn set_ticks(&mut self, last_run: Tick, this_run: Tick)

Overwrite the last_run and this_run tick that are used for change detection.

This is an advanced feature. Refs are usually created by engine-internal code and consumed by end-user code.