foliage

Struct ReflectComponent

pub struct ReflectComponent(/* private fields */);

A struct used to operate on reflected Component trait of a type.

A ReflectComponent for type T can be obtained via [bevy_reflect::TypeRegistration::data].

Implementations

impl ReflectComponent

pub fn insert( &self, entity: &mut EntityWorldMut<'_>, component: &(dyn PartialReflect + 'static), registry: &TypeRegistry, )

Insert a reflected Component into the entity like insert().

pub fn apply<'a>( &self, entity: impl Into<EntityMut<'a>>, component: &(dyn PartialReflect + 'static), )

Uses reflection to set the value of this Component type in the entity to the given value.

Panics

Panics if there is no Component of the given type.

Will also panic if Component is immutable.

pub fn apply_or_insert_mapped( &self, entity: &mut EntityWorldMut<'_>, component: &(dyn PartialReflect + 'static), registry: &TypeRegistry, map: &mut dyn EntityMapper, relationship_hook_mode: RelationshipHookMode, )

Uses reflection to set the value of this Component type in the entity to the given value or insert a new one if it does not exist.

Panics

Panics if Component is immutable.

pub fn remove(&self, entity: &mut EntityWorldMut<'_>)

Removes this Component type from the entity. Does nothing if it doesn’t exist.

pub fn take(&self, entity: &mut EntityWorldMut<'_>) -> Option<Box<dyn Reflect>>

Removes this Component from the entity and returns its previous value.

pub fn contains<'w, 's>( &self, entity: impl Into<FilteredEntityRef<'w, 's>>, ) -> bool

Returns whether entity contains this Component

pub fn reflect<'w, 's>( &self, entity: impl Into<FilteredEntityRef<'w, 's>>, ) -> Option<&'w (dyn Reflect + 'static)>

Gets the value of this Component type from the entity as a reflected reference.

pub fn reflect_mut<'w, 's>( &self, entity: impl Into<FilteredEntityMut<'w, 's>>, ) -> Option<Mut<'w, dyn Reflect>>

Gets the value of this Component type from the entity as a mutable reflected reference.

Panics

Panics if Component is immutable.

pub unsafe fn reflect_unchecked_mut<'a>( &self, entity: UnsafeEntityCell<'a>, ) -> Option<Mut<'a, dyn Reflect>>

Safety

This method does not prevent you from having two mutable pointers to the same data, violating Rust’s aliasing rules. To avoid this:

Only call this method with a UnsafeEntityCell that may be used to mutably access the component on the entity entity

Don’t call this method more than once in the same scope for a given Component.

Panics

Panics if Component is immutable.

pub fn copy( &self, source_world: &World, destination_world: &mut World, source_entity: Entity, destination_entity: Entity, registry: &TypeRegistry, )

Gets the value of this Component type from entity from source_world and applies it to the value of this Component type in entity in destination_world.

Panics

Panics if there is no Component of the given type or either entity does not exist.

pub fn register_component(&self, world: &mut World) -> ComponentId

Register the type of this Component in World, returning its ComponentId.

pub fn new(fns: ReflectComponentFns) -> ReflectComponent

Create a custom implementation of ReflectComponent.

This is an advanced feature, useful for scripting implementations, that should not be used by most users unless you know what you are doing.

Usually you should derive [Reflect] and add the #[reflect(Component)] component to generate a ReflectComponent implementation automatically.

See ReflectComponentFns for more information.

pub fn fn_pointers(&self) -> &ReflectComponentFns

The underlying function pointers implementing methods on ReflectComponent.

This is useful when you want to keep track locally of an individual function pointer.

Calling TypeRegistry::get followed by TypeRegistration::data::<ReflectComponent> can be costly if done several times per frame. Consider cloning ReflectComponent and keeping it between frames, cloning a ReflectComponent is very cheap.

If you only need a subset of the methods on ReflectComponent, use fn_pointers to get the underlying ReflectComponentFns and copy the subset of function pointers you care about.

pub fn map_entities( &self, component: &mut (dyn Reflect + 'static), func: &mut dyn EntityMapper, )

Calls a dynamic version of Component::map_entities.