foliage

Struct Schedules

pub struct Schedules {
    pub ignored_scheduling_ambiguities: BTreeSet<ComponentId>,
    /* private fields */
}

Resource that stores Schedules mapped to ScheduleLabels excluding the current running Schedule.

Fields
ignored_scheduling_ambiguities: BTreeSet<ComponentId>

List of ComponentIds to ignore when reporting system order ambiguity conflicts

Implementations

impl Schedules

pub fn new() -> Schedules

Constructs an empty Schedules with zero initial capacity.

pub fn insert(&mut self, schedule: Schedule) -> Option<Schedule>

Inserts a labeled schedule into the map.

If the map already had an entry for label, schedule is inserted, and the old schedule is returned. Otherwise, None is returned.

pub fn reinsert(&mut self, schedule: Schedule) -> Option<Schedule>

Inserts a labeled schedule into the map.

If the map already had an entry for label, schedule is inserted, and the old schedule is returned. Otherwise, None is returned.

pub fn remove(&mut self, label: impl ScheduleLabel) -> Option<Schedule>

Removes the schedule corresponding to the label from the map, returning it if it existed.

pub fn remove_temporarily( &mut self, label: impl ScheduleLabel, ) -> Option<Schedule>

Removes the schedule corresponding to the label from the map, returning it if it existed, tracks.

pub fn remove_entry( &mut self, label: impl ScheduleLabel, ) -> Option<(Interned<dyn ScheduleLabel>, Schedule)>

Removes the (schedule, label) pair corresponding to the label from the map, returning it if it existed.

pub fn get_temporarily_removed(&self) -> HashSet<Interned<dyn ScheduleLabel>>

Gets a set of temporarily removed schedules

pub fn get_empty_labels(&self) -> HashSet<Interned<dyn ScheduleLabel>>

Gets a set of empty schedule labels

pub fn contains(&self, label: impl ScheduleLabel) -> bool

Does a schedule with the provided label already exist?

pub fn get(&self, label: impl ScheduleLabel) -> Option<&Schedule>

Returns a reference to the schedule associated with label, if it exists.

pub fn get_mut(&mut self, label: impl ScheduleLabel) -> Option<&mut Schedule>

Returns a mutable reference to the schedule associated with label, if it exists.

pub fn entry(&mut self, label: impl ScheduleLabel) -> &mut Schedule

Returns a mutable reference to the schedules associated with label, creating one if it doesn’t already exist.

pub fn iter( &self, ) -> impl Iterator<Item = (&(dyn ScheduleLabel + 'static), &Schedule)>

Returns an iterator over all schedules. Iteration order is undefined.

pub fn iter_mut( &mut self, ) -> impl Iterator<Item = (&(dyn ScheduleLabel + 'static), &mut Schedule)>

Returns an iterator over mutable references to all schedules. Iteration order is undefined.

pub fn configure_schedules( &mut self, schedule_build_settings: ScheduleBuildSettings, )

Applies the provided ScheduleBuildSettings to all schedules.

This mutates all currently present schedules, but does not apply to schedules added in the future.

pub fn allow_ambiguous_component<T>(&mut self, world: &mut World)
where T: Component,

Ignore system order ambiguities caused by conflicts on Components of type T.

pub fn allow_ambiguous_resource<T>(&mut self, world: &mut World)
where T: Resource,

Ignore system order ambiguities caused by conflicts on Resources of type T.

pub fn iter_ignored_ambiguities(&self) -> impl Iterator<Item = &ComponentId>

Iterate through the ComponentId’s that will be ignored.

pub fn print_ignored_ambiguities(&self, components: &Components)

Prints the names of the components and resources with info

May panic or retrieve incorrect names if Components is not from the same world

pub fn add_systems<M>( &mut self, schedule: impl ScheduleLabel, systems: impl IntoScheduleConfigs<Box<dyn System<Out = (), In = ()>>, M>, ) -> &mut Schedules

Adds one or more systems to the Schedule matching the provided ScheduleLabel.

pub fn remove_systems_in_set<M>( &mut self, schedule: impl ScheduleLabel, set: impl IntoSystemSet<M>, world: &mut World, policy: ScheduleCleanupPolicy, ) -> Result<usize, ScheduleError>

Removes all systems in a SystemSet. This will cause the schedule to be rebuilt when the schedule is run again. A ScheduleError is returned if the schedule needs to be Schedule::initialize’d or the set is not found.

pub fn configure_sets<M>( &mut self, schedule: impl ScheduleLabel, sets: impl IntoScheduleConfigs<Interned<dyn SystemSet>, M>, ) -> &mut Schedules

Configures a collection of system sets in the provided schedule, adding any sets that do not exist.

pub fn ignore_ambiguity<M1, M2, S1, S2>( &mut self, schedule: impl ScheduleLabel, a: S1, b: S2, ) -> &mut Schedules
where S1: IntoSystemSet<M1>, S2: IntoSystemSet<M2>,

Suppress warnings and errors that would result from systems in these sets having ambiguities (conflicting access but indeterminate order) with systems in set.

When possible, do this directly in the .add_systems(Update, a.ambiguous_with(b)) call. However, sometimes two independent plugins A and B are reported as ambiguous, which you can only suppress as the consumer of both.