foliage

Trait IntoSystem

pub trait IntoSystem<In, Out, Marker>: Sized
where
    In: SystemInput,
{
    type System: System<In = In, Out = Out>;

    // Required method
    fn into_system(this: Self) -> Self::System;

    // Provided methods
    fn pipe<B, BIn, BOut, MarkerB>(self, system: B) -> IntoPipeSystem<Self, B>
       where Out: 'static,
             B: IntoSystem<BIn, BOut, MarkerB>,
             BIn: for<'a> SystemInput<Inner<'a> = Out> { ... }
    fn map<T, F>(self, f: F) -> IntoAdapterSystem<F, Self>
       where F: Send + Sync + 'static + FnMut(Out) -> T { ... }
    fn with_input<T>(self, value: T) -> WithInputWrapper<Self::System, T>
       where In: for<'i> SystemInput<Inner<'i> = &'i mut T>,
             T: Send + Sync + 'static { ... }
    fn with_input_from<T>(self) -> WithInputFromWrapper<Self::System, T>
       where In: for<'i> SystemInput<Inner<'i> = &'i mut T>,
             T: FromWorld + Send + Sync + 'static { ... }
    fn system_type_id(&self) -> TypeId { ... }
}

Conversion trait to turn something into a System.

Use this to get a system from a function. Also note that every system implements this trait as well.

Usage notes

This trait should only be used as a bound for trait implementations or as an argument to a function. If a system needs to be returned from a function or stored somewhere, use System instead of this trait.

Examples

use bevy_ecs::prelude::*;

fn my_system_function(a_usize_local: Local<usize>) {}

let system = IntoSystem::into_system(my_system_function);

Required Associated Types

type System: System<In = In, Out = Out>

The type of System that this instance converts into.

Required Methods

fn into_system(this: Self) -> Self::System

Turns this value into its corresponding System.

Provided Methods

fn pipe<B, BIn, BOut, MarkerB>(self, system: B) -> IntoPipeSystem<Self, B>
where Out: 'static, B: IntoSystem<BIn, BOut, MarkerB>, BIn: for<'a> SystemInput<Inner<'a> = Out>,

Pass the output of this system A into a second system B, creating a new compound system.

The second system must have In<T> as its first parameter, where T is the return type of the first system.

fn map<T, F>(self, f: F) -> IntoAdapterSystem<F, Self>
where F: Send + Sync + 'static + FnMut(Out) -> T,

Pass the output of this system into the passed function f, creating a new system that outputs the value returned from the function.

// Ignores the output of a system that may fail.
schedule.add_systems(my_system.map(drop));

fn my_system(res: Res<T>) -> Result<(), Err> {
    // ...
}

fn with_input<T>(self, value: T) -> WithInputWrapper<Self::System, T>
where In: for<'i> SystemInput<Inner<'i> = &'i mut T>, T: Send + Sync + 'static,

Passes a mutable reference to value as input to the system each run, turning it into a system that takes no input.

Self can have any SystemInput type that takes a mutable reference to T, such as InMut.

Example

fn my_system(InMut(value): InMut<usize>) {
    *value += 1;
    if *value > 10 {
       println!("Value is greater than 10!");
    }
}

schedule.add_systems(my_system.with_input(0));

fn with_input_from<T>(self) -> WithInputFromWrapper<Self::System, T>
where In: for<'i> SystemInput<Inner<'i> = &'i mut T>, T: FromWorld + Send + Sync + 'static,

Passes a mutable reference to a value of type T created via FromWorld as input to the system each run, turning it into a system that takes no input.

Self can have any SystemInput type that takes a mutable reference to T, such as InMut.

Example

struct MyData {
    value: usize,
}

impl FromWorld for MyData {
    fn from_world(world: &mut World) -> Self {
        // Fetch from the world the data needed to create `MyData`
    }
}

fn my_system(InMut(data): InMut<MyData>) {
    data.value += 1;
    if data.value > 10 {
        println!("Value is greater than 10!");
    }
}
schedule.add_systems(my_system.with_input_from::<MyData>());

fn system_type_id(&self) -> TypeId

Get the TypeId of the System produced after calling into_system.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.