foliage

Trait IntoSystemSet

pub trait IntoSystemSet<Marker>: Sized {
    type Set: SystemSet;

    // Required method
    fn into_system_set(self) -> Self::Set;
}

Types that can be converted into a SystemSet.

Usage notes

This trait should only be used as a bound for trait implementations or as an argument to a function. If a system set needs to be returned from a function or stored somewhere, use SystemSet instead of this trait.

Required Associated Types

type Set: SystemSet

The type of SystemSet this instance converts into.

Required Methods

fn into_system_set(self) -> Self::Set

Converts this instance to its associated SystemSet type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.