foliage

Trait ReadOnlySystem

pub unsafe trait ReadOnlySystem: System {
    // Provided method
    fn run_readonly(
        &mut self,
        input: <Self::In as SystemInput>::Inner<'_>,
        world: &World,
    ) -> Result<Self::Out, RunSystemError> { ... }
}

System types that do not modify the World when run. This is implemented for any systems whose parameters all implement ReadOnlySystemParam.

Note that systems which perform deferred mutations (such as with Commands) may implement this trait.

Safety

This must only be implemented for system types which do not mutate the World when System::run_unsafe is called.

Provided Methods

fn run_readonly( &mut self, input: <Self::In as SystemInput>::Inner<'_>, world: &World, ) -> Result<Self::Out, RunSystemError>

Runs this system with the given input in the world.

Unlike System::run, this can be called with a shared reference to the world, since this system is known not to modify the world.