foliage

Trait System

pub trait System:
    Send
    + Sync
    + 'static {
    type In: SystemInput;
    type Out;

    // Required methods
    fn name(&self) -> DebugName;
    fn flags(&self) -> SystemStateFlags;
    unsafe fn run_unsafe(
        &mut self,
        input: <Self::In as SystemInput>::Inner<'_>,
        world: UnsafeWorldCell<'_>,
    ) -> Result<Self::Out, RunSystemError>;
    fn apply_deferred(&mut self, world: &mut World);
    fn queue_deferred(&mut self, world: DeferredWorld<'_>);
    fn initialize(&mut self, _world: &mut World) -> FilteredAccessSet;
    fn check_change_tick(&mut self, check: CheckChangeTicks);
    fn get_last_run(&self) -> Tick;
    fn set_last_run(&mut self, last_run: Tick);

    // Provided methods
    fn system_type(&self) -> TypeId { ... }
    fn type_id(&self) -> TypeId { ... }
    fn is_send(&self) -> bool { ... }
    fn is_exclusive(&self) -> bool { ... }
    fn has_deferred(&self) -> bool { ... }
    fn run(
        &mut self,
        input: <Self::In as SystemInput>::Inner<'_>,
        world: &mut World,
    ) -> Result<Self::Out, RunSystemError> { ... }
    fn run_without_applying_deferred(
        &mut self,
        input: <Self::In as SystemInput>::Inner<'_>,
        world: &mut World,
    ) -> Result<Self::Out, RunSystemError> { ... }
    fn default_system_sets(&self) -> Vec<Interned<dyn SystemSet>> { ... }

}

An ECS system that can be added to a Schedule

Systems are functions with all arguments implementing SystemParam.

Systems are added to an application using App::add_systems(Update, my_system) or similar methods, and will generally run once per pass of the main loop.

Systems are executed in parallel, in opportunistic order; data access is managed automatically. It’s possible to specify explicit execution order between specific systems, see IntoScheduleConfigs.

Required Associated Types

type In: SystemInput

The system’s input.

type Out

The system’s output.

Required Methods

fn name(&self) -> DebugName

Returns the system’s name.

fn flags(&self) -> SystemStateFlags

Returns the SystemStateFlags of the system.

unsafe fn run_unsafe( &mut self, input: <Self::In as SystemInput>::Inner<'_>, world: UnsafeWorldCell<'_>, ) -> Result<Self::Out, RunSystemError>

Runs the system with the given input in the world. Unlike System::run, this function can be called in parallel with other systems and may break Rust’s aliasing rules if used incorrectly, making it unsafe to call.

Unlike System::run, this will not apply deferred parameters, which must be independently applied by calling System::apply_deferred at later point in time.

Safety

The caller must ensure that world has permission to access any world data registered in the access returned from System::initialize. There must be no conflicting simultaneous accesses while the system is running.

If System::is_exclusive returns true, then it must be valid to call UnsafeWorldCell::world_mut on world.

fn apply_deferred(&mut self, world: &mut World)

Applies any Deferred system parameters (or other system buffers) of this system to the world.

This is where Commands get applied.

fn queue_deferred(&mut self, world: DeferredWorld<'_>)

Enqueues any Deferred system parameters (or other system buffers) of this system into the world’s command buffer.

fn initialize(&mut self, _world: &mut World) -> FilteredAccessSet

Initialize the system.

Returns a FilteredAccessSet with the access required to run the system.

fn check_change_tick(&mut self, check: CheckChangeTicks)

Checks any Ticks stored on this system and wraps their value if they get too old.

This method must be called periodically to ensure that change detection behaves correctly. When using bevy’s default configuration, this will be called for you as needed.

fn get_last_run(&self) -> Tick

Gets the tick indicating the last time this system ran.

fn set_last_run(&mut self, last_run: Tick)

Overwrites the tick indicating the last time this system ran.

Warning

This is a complex and error-prone operation, that can have unexpected consequences on any system relying on this code. However, it can be an essential escape hatch when, for example, you are trying to synchronize representations using change detection and need to avoid infinite recursion.

Provided Methods

fn system_type(&self) -> TypeId

Returns the TypeId of the underlying system type.

fn type_id(&self) -> TypeId

👎Deprecated since 0.19.0:
Use system_type instead. This method shadows Any::type_id and will be removed in a future release.

Returns the TypeId of the underlying system type.

Use system_type instead.

fn is_send(&self) -> bool

Returns true if the system is Send.

fn is_exclusive(&self) -> bool

Returns true if the system must be run exclusively.

fn has_deferred(&self) -> bool

Returns true if system has deferred buffers.

fn run( &mut self, input: <Self::In as SystemInput>::Inner<'_>, world: &mut World, ) -> Result<Self::Out, RunSystemError>

Runs the system with the given input in the world.

For read-only systems, see run_readonly, which can be called using &World.

Unlike System::run_unsafe, this will apply deferred parameters immediately.

fn run_without_applying_deferred( &mut self, input: <Self::In as SystemInput>::Inner<'_>, world: &mut World, ) -> Result<Self::Out, RunSystemError>

Runs the system with the given input in the world.

fn default_system_sets(&self) -> Vec<Interned<dyn SystemSet>>

Returns the system’s default system sets.

Each system will create a default system set that contains the system.