Struct foliage::wgpu::Device

pub struct Device { /* private fields */ }

Open connection to a graphics and/or compute device.

Responsible for the creation of most rendering and compute resources. These are then used in commands, which are submitted to a Queue.

A device may be requested from an adapter with Adapter::request_device.

Corresponds to WebGPU GPUDevice.

Implementations

impl Device

pub fn poll(&self, maintain: Maintain<SubmissionIndex>) -> MaintainResult

Check for resource cleanups and mapping callbacks. Will block if Maintain::Wait is passed.

Return true if the queue is empty, or false if there are more queue submissions still in flight. (Note that, unless access to the Queue is coordinated somehow, this information could be out of date by the time the caller receives it. Queues can be shared between threads, so other threads could submit new work at any time.)

When running on WebGPU, this is a no-op. Devices are automatically polled.

pub fn features(&self) -> Features

The features which can be used on this device.

No additional features can be used, even if the underlying adapter can support them.

pub fn limits(&self) -> Limits

The limits which can be used on this device.

No better limits can be used, even if the underlying adapter can support them.

pub fn create_shader_module( &self, desc: ShaderModuleDescriptor<'_> ) -> ShaderModule

Creates a shader module from either SPIR-V or WGSL source code.

pub unsafe fn create_shader_module_unchecked( &self, desc: ShaderModuleDescriptor<'_> ) -> ShaderModule

Creates a shader module from either SPIR-V or WGSL source code without runtime checks.

Safety

In contrast with create_shader_module this function creates a shader module without runtime checks which allows shaders to perform operations which can lead to undefined behavior like indexing out of bounds, thus it’s the caller responsibility to pass a shader which doesn’t perform any of this operations.

This has no effect on web.

pub unsafe fn create_shader_module_spirv( &self, desc: &ShaderModuleDescriptorSpirV<'_> ) -> ShaderModule

Creates a shader module from SPIR-V binary directly.

Safety

This function passes binary data to the backend as-is and can potentially result in a driver crash or bogus behaviour. No attempt is made to ensure that data is valid SPIR-V.

See also include_spirv_raw! and util::make_spirv_raw.

pub fn create_command_encoder( &self, desc: &CommandEncoderDescriptor<Option<&str>> ) -> CommandEncoder

Creates an empty CommandEncoder.

pub fn create_render_bundle_encoder( &self, desc: &RenderBundleEncoderDescriptor<'_> ) -> RenderBundleEncoder<'_>

Creates an empty RenderBundleEncoder.

pub fn create_bind_group(&self, desc: &BindGroupDescriptor<'_>) -> BindGroup

Creates a new BindGroup.

pub fn create_bind_group_layout( &self, desc: &BindGroupLayoutDescriptor<'_> ) -> BindGroupLayout

Creates a BindGroupLayout.

pub fn create_pipeline_layout( &self, desc: &PipelineLayoutDescriptor<'_> ) -> PipelineLayout

Creates a PipelineLayout.

pub fn create_render_pipeline( &self, desc: &RenderPipelineDescriptor<'_> ) -> RenderPipeline

Creates a RenderPipeline.

pub fn create_compute_pipeline( &self, desc: &ComputePipelineDescriptor<'_> ) -> ComputePipeline

Creates a ComputePipeline.

pub fn create_buffer(&self, desc: &BufferDescriptor<Option<&str>>) -> Buffer

Creates a Buffer.

pub fn create_texture( &self, desc: &TextureDescriptor<Option<&str>, &[TextureFormat]> ) -> Texture

Creates a new Texture.

desc specifies the general format of the texture.

pub unsafe fn create_texture_from_hal<A>( &self, hal_texture: <A as Api>::Texture, desc: &TextureDescriptor<Option<&str>, &[TextureFormat]> ) -> Texturewhere A: HalApi,

Creates a Texture from a wgpu-hal Texture.

Safety

hal_texture must be created from this device internal handle

hal_texture must be created respecting desc

hal_texture must be initialized

pub unsafe fn create_buffer_from_hal<A>( &self, hal_buffer: <A as Api>::Buffer, desc: &BufferDescriptor<Option<&str>> ) -> Bufferwhere A: HalApi,

Creates a Buffer from a wgpu-hal Buffer.

Safety

hal_buffer must be created from this device internal handle

hal_buffer must be created respecting desc

hal_buffer must be initialized

pub fn create_sampler(&self, desc: &SamplerDescriptor<'_>) -> Sampler

Creates a new Sampler.

desc specifies the behavior of the sampler.

pub fn create_query_set( &self, desc: &QuerySetDescriptor<Option<&str>> ) -> QuerySet

Creates a new QuerySet.

pub fn on_uncaptured_error( &self, handler: Box<dyn UncapturedErrorHandler<Output = ()>> )

Set a callback for errors that are not handled in error scopes.

pub fn push_error_scope(&self, filter: ErrorFilter)

Push an error scope.

pub fn pop_error_scope( &self ) -> impl Future<Output = Option<Error>> + WasmNotSend

Pop an error scope.

pub fn start_capture(&self)

Starts frame capture.

pub fn stop_capture(&self)

Stops frame capture.

pub unsafe fn as_hal<A, F, R>(&self, hal_device_callback: F) -> Option<R>where A: HalApi, F: FnOnce(Option<&<A as Api>::Device>) -> R,

Apply a callback to this Device’s underlying backend device.

If this Device is implemented by the backend API given by A (Vulkan, Dx12, etc.), then apply hal_device_callback to Some(&device), where device is the underlying backend device type, A::Device.

If this Device uses a different backend, apply hal_device_callback to None.

The device is locked for reading while hal_device_callback runs. If the callback attempts to perform any wgpu operations that require write access to the device (destroying a buffer, say), deadlock will occur. The locks are automatically released when the callback returns.

Safety

The raw handle passed to the callback must not be manually destroyed.

pub fn destroy(&self)

Destroy this device.

pub fn set_device_lost_callback( &self, callback: impl Fn(DeviceLostReason, String) + Send + 'static )

Set a DeviceLostCallback on this device.

impl Device

pub fn global_id(&self) -> Id<Device>

Returns a globally-unique identifier for this Device.

Calling this method multiple times on the same object will always return the same value. The returned value is guaranteed to be different for all resources created from the same Instance.